新增 pick 和 randomColor
