+++ src/utils.js
@@ -2,7 +2,7 @@
  * @Author: Russ Zhong 
  * @Date: 2018-12-11 14:17:11 
  * @Last Modified by: Russ Zhong
- * @Last Modified time: 2018-12-20 08:47:12
+ * @Last Modified time: 2018-12-20 11:16:10
  */
 
 // Functions in this module won't be exposed to users.
@@ -12,6 +12,7 @@ const toString = Object.prototype.toString;
 const slice = Array.prototype.slice;
 const join = Array.prototype.join;
 const hasOwnProp = Object.prototype.hasOwnProperty;
+
 
 const throwTypeErr = function(msg) {
   throw new TypeError(msg);
