+++ test/Util.test.js
@@ -2,7 +2,7 @@
  * @Author: Russ Zhong 
  * @Date: 2018-12-11 11:06:50 
  * @Last Modified by: Russ Zhong
- * @Last Modified time: 2018-12-20 11:06:02
+ * @Last Modified time: 2018-12-20 13:30:44
  */
 
 const expect = require('expect.js');
@@ -32,7 +32,8 @@ const {
   keys,
   has,
   equals,
-  deepClone
+  deepClone,
+  randomInt
 } = require('../src/packages/Util');
 
 describe('*************************************测试工具函数*************************************', function() {
@@ -691,6 +692,26 @@ describe('*************************************测试工具函数***************
       p1.fs.push(12);
       expect(p1.fs).length(4);
       expect(p.fs).length(3);
+    });
+  });
+  describe('测试 randomInt', function() {
+    it('非法参数报错', function() {
+      for (let i = 0; i< 100; i++) {
+        expect(() => {randomInt()}).to.throwError();
+        expect(() => {randomInt(105)}).to.throwError();
+        expect(() => {randomInt(105, 100)}).to.throwError();
+      }
+    });
+    it('返回正确结果', function() {
+      for (let i = 0; i < 100; i++) {
+        expect(randomInt(0, 1)).within(0, 1);
+      }
+      for (let i = 0; i < 100; i++) {
+        expect(randomInt(5, 7)).within(5, 7);
+      }
+      for (let i = 0; i< 100; i++) {
+        expect(randomInt(100, 105)).within(100, 105);
+      }
     });
   });
 });
